Add JSON aggregted data
Noticed some performance issues when subscribing to indiviual values.
Aggregating data on the device end should make things smoother, and
stay within the transaction/hour limit.

This is the initial step, the target is to end up with the DataSource
base class completely handling mbed subscription and value updating,
so the derived classes only have to deal with actually reading from the
hardware.

Signed-off-by: Cristian Prundeanu <cristian.prundeanu@arm.com>